requirements.txt
pdf2image==1.16.0
docx2pdf==0.1.8
labelImg==1.8.6
numpy==1.23.2
pascal-voc==0.0.3
pytesseract==0.3.8
